backend/requirements.txt
python-dotenv==1.0.0
deepgram-sdk==3.5.0
sounddevice==0.4.6
numpy==1.26.4
requests==2.31.0

flask==3.0.0
flask-socketio==5.3.5
flask-cors==4.0.0
python-socketio==5.10.0
#webrtcvad==2.0.10
python-engineio==4.8.0
simple-websocket==1.0.0


Flask-Bcrypt
PyJWT
pymongo